Fix for PT Trace support after prior fix of static branch target
Fixing the static branch target revealed other issues in the pt
frontend, notably:
* the prior fix for ensuring gaps in the trace not caused by CF insns
didn't handle when the deltas between the instructions could not be
encoded in 32 bits. We fixed this by changing the static target in the
generated ctype_pin_inst to the "correct" target for this fake
instruction.
* There are some places in the traces where the instruction executed
after a direct call is not the target of the call. This was fixed much
in the same way as the prior item - the "bad" call/jmp was overwritten
with a jmp to the "real" target, with the same fix for if the delta did
not fit into 32 bits.