Create a wrapper for instruction memreq callbacks.
This allows different things to be done depending on the type of instruction
memreq: E.g. when later we add a uop cache prefetch,
or if we want to do something extra for an FDIP prefetch.
Any two memreqs that are coalesced must have the same callback function since
the newer memreq's callback is ignored.