This commit includes the following changes including new features.
1. 2-phase fully warmed up running
- Enable the warmup mode to collect the counter information where the
instruction cache line is useful or not
- the counter information can be consumed during SIMULATION_MODE
2. DEBUG_FDIP
- Add print statements useful for debugging related to FDIP
3. Collecting the usefulness of cache lines with FDIP_HASH_ENABLE,
FDIP_HASH_USEFUL_WEIGHT, FDIP_HASH_BIAS
- Add two hash tables for useful/unuseful cache lines respectively
- A cache line address is inserted into fetch_cl_addr when it is firstly seen,
and removed from fetch_cl_addr when it is evicted or hit
and counting the numbers with cnt_useful and cnt_unuseful.
- FDIP determines whether to emit a prefetch for a cache line based on the following condition.
unuseful_cnt == 0 || useful_cnt * WEIGHT >= unuseful_cnt + BIAS
4. Fix for a bug (need to discuss with Scarab folks)
- symptom : There are missing cases for merging opportunity of a new memory request
- RC : The existing memory mechanism iterates all the queues when it finds a matching request
instead of the whole memory request buffer even if there is a period where a request is not in any queue.
- fix : Find a valid memory request directly from the whole req_buffer and merging it
directly into the buffer instead of finding one from the queues.