Simplify the modeling of the icache/uop cache latencies.
The goal is to decrease code complexity to minimize risk of errors/bugs,
and verify that results are the same.
This is equivalent to the previous method using IC_WAIT_FOR_TIMER,
but has less functionality (does not support UOP_CACHE_LATENCY > 1, and
is less intuitive since icache fetch latency is modeled by decode stage length.
After verifying that this is equivalent to the current implementation, I can
discard this commit.

Since ops from the uop cache do not go through decode, an equivalent way
of simulating icache latency greater than one is to increase the decode
latency by that amount. This will give identical performance.

We can assume that accessing the uop cache is always one cycle since
it is a small near-core cache, so we remove that parameter.