Bugfix: Do not use op->table_info->btb_miss to identify if an op is a branch that missed in the BTB.
op->table_info->btb_miss is sometimes set to TRUE for non control flow
changing instructions. To identify if an op is a branch that missed in the
BTB, first check if op->table_info->cf_type.
This check was used to decide when to break instruction fetch. This bug was
causing some ops from the uop cache to incorrectly go to the decode stage.