Create mechanism to input a list of PWs to treat as priority (with REPL_STICKY_PRIORITY_LINES).
Input is a CSV file with no header and the second column being a list of PWs sorted
from most to least important. The number of PWs to treat as priority
is set by the NUM_PRIORITY_PWS parameter. The path to the CSV is provided
with PW_PRIORITY_LIST_FILEPATH.