Improve correctness of branch predictor by implementing the following approach: 1. Set oracle_info.pred and oracle_info.pred_npc based on the pred_func even on BTB miss. This is compute the branch predictor outcome that would either happen at fetch (BTB hit) or decode (BTB miss) 2. Utilize the branch predictor outcome to compute recover_at_decode, recover_at_exec 3. Overwrite oracle_info.pred/oracle_info.pred_npc with NOT_TAKEN/pc_plus_offset in the case of an BTB miss
This assumes that all pipe stages that consume pred/pred_npc are between fetch and decode and hence they need to consider BTB misses for prediction.
The actual prediction performed at decode is no longer available between fetch and decode. If this is required we need to add new fields for
decode_pred and decode_pred_npc which are not overwritten by a BTB miss.