Bugfix for PRIORITIZE_PWS_AFTER_FIRST_STICKY_UNTIL_BACKEND_STALL
make_accesses_priority should be set beginning on the first ACCESS of a top PW,
not on its insert -- this misses all the PWs that are accessed
between the top PW access and its insertion!
All accessed PWs until the next backend stall need to be made priority.

Ensure PWs that follow a top PW is inserted or ACCESSED are sticky.
Previously this was toggled only on an insert of a top PW.
Then for insert, we can just see if the PW began accumulation while
make_accesses_priority was toggled.

* Move to stat_event_new_pw_accessed to uop_cache, since that's a more logical place for it.
* Add a couple uop cache debug statements