Add option UOC_IC_SWITCH_FRAG_DISABLE to disable fragmentation caused when switching between uop cache / icache and vice versa.
This option allows the uop cache and icache to fetch in the same cycle.
The map stage then recombines these two inputs in the same cycle.
This enables full issue width to be used despite switching, similar to
FETCH_ACROSS_CACHE_LINES, but across uop cache and icache lines.

In the process, change back map_stage so that it is only updated once per cycle.