bug fixes after the frontend upgrade
Fix 1: The npc of the last intruction in the trace cannot be set by its following non-existing instruction. However, the way it was set previously triggered a frontend assertion. It is fixed by setting the npc to itself.
Fix 2: Previously the pid and tid were initialized only if the ins_id is zero. However, if the scarab fast-forward was used, ins_id will not be zero. As a result all instructions were skipped. It is fixed by changing the condition.